requirements.txt
Appium_Python_Client==2.6.0
pytest==7.1.2